Tüm Controllerlarda silme işlemi kontrolü gerçekleştirildi.

+++ site/panel/application/controllers/Email.php
@@ -179,6 +179,9 @@ class Email extends MY_Controller
         }
     }
     public function delete($id){
+        if (!isAllowedDeleteModule($this->router->fetch_class())):
+            redirect(base_url($this->router->fetch_class()));
+        endif;
         
         $delete = $this->email_model->delete(
           array(
