+++ site/panel/application/controllers/Popup.php
@@ -194,6 +194,9 @@ class Popup extends MY_Controller
     }
 
     public function delete($id){
+        if (!isAllowedDeleteModule($this->router->fetch_class())):
+            redirect(base_url($this->router->fetch_class()));
+        endif;
 
             $delete = $this->popup_model->delete(
                 array(
