+++ site/panel/application/controllers/Users.php
@@ -333,7 +333,10 @@ class Users extends MY_Controller
     }
 
     public function delete($id){
-        
+        if (!isAllowedDeleteModule($this->router->fetch_class())):
+            redirect(base_url($this->router->fetch_class()));
+        endif;
+
         $delete = $this->user_model->delete(
           array(
               "id" => $id
